Correct type of occupancy grid data

+++ tango_ros_common/tango_ros_native/include/tango_ros_native/tango_ros_conversions_helper.h
@@ -28,9 +28,9 @@
 
 namespace tango_ros_conversions_helper {
 const int NUMBER_OF_FIELDS_IN_POINT_CLOUD = 4;
-constexpr int FREE_CELL = 0;
-constexpr int OCCUPIED_CELL = 100;
-constexpr int UNKNOWN_CELL = -1;
+constexpr int8_t FREE_CELL = 0;
+constexpr int8_t OCCUPIED_CELL = 100;
+constexpr int8_t UNKNOWN_CELL = -1;
 
 // Converts a TangoPoseData to a geometry_msgs::TransformStamped.
 // @param pose, TangoPoseData to convert.
